feat: Add Vehiculo and Accesorio classes
This commit adds the Vehiculo and Accesorio classes to the CleanArchitecture.Domain.Vehiculos namespace. The Vehiculo class is a sealed class representing a vehicle, with properties such as Id, Modelo, Vin, and various location details. The Accesorio enum represents different accessories that can be associated with a vehicle, such as wifi, AireAcondicionado, AppleCar, AndroidCar, and Mapas.