feat: Add Entity class
This commit adds the Entity class to the CleanArchitecture.Domain.Abstractions namespace. The Entity class is an abstract class representing an entity in the domain, with a protected constructor that takes an id parameter and a public property for the id. This class serves as a base class for other domain entities.